f2

+++ tasks/seq/gordeeva_t_shell_sort_batcher_merge/src/ops_seq.cpp
@@ -57,7 +57,7 @@ bool gordeeva_t_shell_sort_batcher_merge_seq::TestTaskSequential::run() {
 
   shellSort(input_, input_.size());
 
-  for (int i = 0; i < input_.size(); i++) {
+  for (size_t i = 0; i < input_.size(); i++) {
     res_[i] = input_[i];
   }
